+++ ezdxf/options.py
@@ -16,7 +16,12 @@ class Options(object):
         self.template_dir = None
         self.logger = logging.getLogger(LOGGER_NAME)
         self.debug = False  # important to set after setup of self.logger
-        self.compress_default_chunks = False  # compress tags of unknown chunks of tgas like unknown sections
+
+        # compress tags of unknown chunks of tags like unknown sections
+        self.compress_default_chunks = False
+
+        # compress binary data tags 310-319, collects multiple succeeding binary tags as one compressed tag
+        self.compress_binary_data = False
 
     @property
     def debug(self):